+++ plugin/eth.go
@@ -12,66 +12,78 @@ import (
 	eth "github.com/ipfs/go-ipld-eth"
 )
 
+// Plugins declare what and how many of these will be defined.
 var Plugins = []plugin.Plugin{
 	&EthereumPlugin{},
 }
 
+// EthereumPlugin is the main structure.
 type EthereumPlugin struct{}
 
 var _ plugin.PluginIPLD = (*EthereumPlugin)(nil)
 
+// Init complies with the plugin.Plugin interface.
+// Use node.BlockDecoder.Register() instead.
+// see https://github.com/ipfs/go-ipld-eth/issues/1#issuecomment-316777885
 func (ep *EthereumPlugin) Init() error {
 	return nil
 }
 
+// Name returns the name of this plugin.
 func (ep *EthereumPlugin) Name() string {
 	return "ipld-ethereum"
 }
 
+// Version returns the version of this plugin.
 func (ep *EthereumPlugin) Version() string {
-	return "0.0.2"
+	return "0.0.3"
 }
 
+/*
+  INPUT
+*/
+
+// RegisterInputEncParsers enters the encode parsers needed to put the blocks into IPLD.
+func (ep *EthereumPlugin) RegisterInputEncParsers(iec coredag.InputEncParsers) error {
+	iec.AddParser("raw", "eth-block", EthBlockInputParser)
+	return nil
+}
+
+// EthBlockInputParser will take the piped input,
+// an RLP binary of a block header, to return an IPLD node slice.
+func EthBlockInputParser(r io.Reader) ([]node.Node, error) {
+	blockHeader, err := eth.FromBlockHeaderRLP(r)
+	if err != nil {
+		return nil, err
+	}
+	var out []node.Node
+	out = append(out, blockHeader)
+	return out, nil
+}
+
+/*
+  OUTPUT
+*/
+
+// RegisterBlockDecoders enters which functions will help us to decode the requested IPLD blocks.
 func (ep *EthereumPlugin) RegisterBlockDecoders(dec node.BlockDecoder) error {
-	dec.Register(eth.MEthBlock, EthBlockParser)
+	dec.Register(eth.MEthBlock, EthBlockParser) // eth-block
 	dec.Register(eth.MEthTx, EthTxParser)
 	dec.Register(eth.MEthTxTrie, EthTxTrieParser)
 	return nil
 }
 
-func (ep *EthereumPlugin) RegisterInputEncParsers(iec coredag.InputEncParsers) error {
-	iec.AddParser("raw", "eth", BlockInputParser)
-	return nil
-}
-
+// EthBlockParser takes care of the eth-block IPLD objects (ethereum block headers)
 func EthBlockParser(b block.Block) (node.Node, error) {
 	return eth.DecodeBlock(bytes.NewReader(b.RawData()))
 }
 
+// EthTxParser takes care of the eth-tx IPLD objects (ethereum transactions)
 func EthTxParser(b block.Block) (node.Node, error) {
 	return eth.ParseTx(b.RawData())
 }
 
+// EthTxTrieParser takes care of the eth-tx-trie objects (ethereum transaction trie nodes)
 func EthTxTrieParser(b block.Block) (node.Node, error) {
 	return eth.NewTrieNode(b.RawData())
-}
-
-func BlockInputParser(r io.Reader) ([]node.Node, error) {
-	blk, txs, tries, uncles, err := eth.FromRlpBlockMessage(r)
-	if err != nil {
-		return nil, err
-	}
-
-	var out []node.Node
-	out = append(out, blk)
-	for _, tx := range txs {
-		out = append(out, tx)
-	}
-	for _, t := range tries {
-		out = append(out, t)
-	}
-	for _, unc := range uncles {
-		out = append(out, unc)
-	}
-	return out, nil
 }