+++ plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## `0.0.3`
+  * Improve in-code docs for the plugin loader.
+  * Implement `eth-block` (codec `0x90`), as a representation of an ethereum block header.
+
 ## `0.0.2`
 
 * First Implementation of vetting tools: